requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.10
ChatbotAPI==1.4.0
discord.ext.context==0.1.8
discord.py==1.7.3
idna==3.3
multidict==5.2.0
pyjokes==0.6.0
python-dotenv==0.19.2
requests==2.27.1
typing_extensions==4.0.1
urllib3==1.26.8
yarl==1.7.2
absl-py==1.0.0
aiohttp==3.7.4.post0
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.4.0
beautifulsoup4==4.10.0
cachetools==5.0.0
certifi==2021.10.8
chardet==4.0.0
deepface==0.0.74
charset-normalizer==2.0.12
colorama==0.4.4
discord==1.7.3
discord.py==1.7.3
filelock==3.6.0
flatbuffers==2.0
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.11.3
keras==2.8.0
Keras-Preprocessing==1.1.2
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==2.1.1
multidict==6.0.2
numpy==1.22.3
oauthlib==3.2.0
opt-einsum==3.3.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PySocks==1.7.1
python-dotenv==0.19.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
soupsieve==2.3.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tqdm==4.63.0
typing_extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
wrapt==1.14.0
yarl==1.7.2
zipp==3.7.0
